revise test case for competition
