+++ phpunit.xml
@@ -1,7 +1,7 @@
 <phpunit bootstrap="./test/bootstrap.php">
     <testsuites>
-      <testsuite name="provider">
-        <file>./test/Unit/ProviderTest.php</file>
+      <testsuite name="competition">
+        <file>./test/Unit/CompetitionApiTest.php</file>
       </testsuite>
   </testsuites>
 </phpunit>
